[shadowmap] feat: add server and web UI
