+++ Cargo.toml
@@ -62,6 +62,9 @@ chrono = "0.4.41"
 dialoguer = "0.12.0"
 once_cell = "1.21.3"
 papaya = "0.2.3"
+actix-web = "4"
+sqlx = { version = "0.7", features = ["runtime-tokio-rustls", "sqlite"] }
+uuid = { version = "1", features = ["serde", "v4"] }
 
 [dev-dependencies]
 tokio-test = "0.4"
@@ -72,3 +75,7 @@ lto = true
 codegen-units = 1
 panic = "abort"
 strip = true
+
+[[bin]]
+name = "shadowmap-server"
+path = "src/server.rs"
