+++ web/index.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html>
+<head>
+<title>ShadowMap Dashboard</title>
+<script>
+const API_URL = 'http://localhost:8080';
+
+async function submitJob() {
+  const domain = document.getElementById('domain').value;
+  const token = document.getElementById('token').value;
+  const res = await fetch(API_URL + '/jobs', {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': 'Bearer ' + token
+    },
+    body: JSON.stringify({domain})
+  });
+  const data = await res.json();
+  document.getElementById('jobId').innerText = data.id;
+}
+
+async function checkStatus() {
+  const id = document.getElementById('statusId').value;
+  const token = document.getElementById('token').value;
+  const res = await fetch(API_URL + '/jobs/' + id, {
+    headers: { 'Authorization': 'Bearer ' + token }
+  });
+  const data = await res.json();
+  document.getElementById('statusResult').innerText = data.status;
+}
+
+async function downloadReport() {
+  const id = document.getElementById('statusId').value;
+  const token = document.getElementById('token').value;
+  const res = await fetch(API_URL + '/jobs/' + id + '/report', {
+    headers: { 'Authorization': 'Bearer ' + token }
+  });
+  const txt = await res.text();
+  document.getElementById('report').textContent = txt;
+}
+</script>
+</head>
+<body>
+<h1>ShadowMap</h1>
+<div>
+Token: <input id="token" type="text" value="testtoken"><br/>
+Domain: <input id="domain" type="text"><button onclick="submitJob()">Submit</button>
+<p>Job ID: <span id="jobId"></span></p>
+</div>
+<div>
+Job ID: <input id="statusId" type="text"><button onclick="checkStatus()">Check Status</button><button onclick="downloadReport()">Download Report</button>
+<p>Status: <span id="statusResult"></span></p>
+<pre id="report"></pre>
+</div>
+</body>
+</html>
